# api-specification.yaml
openapi: 3.0.3
info:
  title: CinemaAbyss API
  description: |
    API спецификация для системы CinemaAbyss, включающая монолит и микросервисы.
    
    Система CinemaAbyss представляет собой платформу для управления фильмами, пользователями, платежами и подписками.
    Архитектура системы включает монолитное приложение и выделенные микросервисы.
  version: 1.0.0
  contact:
    name: CinemaAbyss Team
    email: support@cinemaabyss.com
servers:
  - url: http://localhost:8000
    description: API Gateway (Proxy Service)
  - url: http://localhost:8080
    description: Монолитное приложение
  - url: http://localhost:8081
    description: Микросервис фильмов
  - url: http://localhost:8082
    description: Микросервис событий
...
paths:
...
  /api/events/movie:
    post:
      tags:
        - events
      summary: Создание события фильма
      description: Регистрирует новое событие, связанное с фильмом
      operationId: createMovieEvent
      requestBody:
        description: Данные события фильма
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieEvent'
      responses:
        '201':
          description: Событие успешно создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/events/user:
    post:
      tags:
        - events
      summary: Создание события пользователя
      description: Регистрирует новое событие, связанное с пользователем
      operationId: createUserEvent
      requestBody:
        description: Данные события пользователя
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserEvent'
      responses:
        '201':
          description: Событие успешно создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/events/payment:
    post:
      tags:
        - events
      summary: Создание события платежа
      description: Регистрирует новое событие, связанное с платежом
      operationId: createPaymentEvent
      requestBody:
        description: Данные события платежа
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentEvent'
      responses:
        '201':
          description: Событие успешно создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
...
    MovieEvent:
      type: object
      properties:
        movie_id:
          type: integer
          description: Идентификатор фильма
          example: 1
        title:
          type: string
          description: Название фильма
          example: "Inception"
        action:
          type: string
          description: Действие с фильмом
          example: "viewed"
        user_id:
          type: integer
          description: Идентификатор пользователя (опционально)
          example: 1
        rating:
          type: number
          format: float
          description: Рейтинг (опционально)
          example: 8.5
        genres:
          type: array
          description: Жанры фильма (опционально)
          items:
            type: string
          example: ["Sci-Fi", "Action"]
        description:
          type: string
          description: Описание фильма (опционально)
          example: "A mind-bending thriller"
      required:
        - movie_id
        - title
        - action
    UserEvent:
      type: object
      properties:
        user_id:
          type: integer
          description: Идентификатор пользователя
          example: 1
        username:
          type: string
          description: Имя пользователя (опционально)
          example: "john_doe"
        email:
          type: string
          description: Email пользователя (опционально)
          example: "john.doe@example.com"
        action:
          type: string
          description: Действие пользователя
          example: "registered"
        occurred_at:
          type: string
          format: date-time
          description: Время создания события сервисом
          example: "2026-07-27T10:00:00Z"
      required:
        - user_id
        - action
        - timestamp
    PaymentEvent:
      type: object
      properties:
        payment_id:
          type: integer
          description: Идентификатор платежа
          example: 1
        user_id:
          type: integer
          description: Идентификатор пользователя
          example: 1
        amount:
          type: number
          format: float
          description: Сумма платежа
          example: 9.99
        status:
          type: string
          description: Статус платежа
          example: "completed"
        timestamp:
          type: string
          format: date-time
          description: Время платежа
          example: "2023-01-15T14:30:00Z"
        method_type:
          type: string
          description: Тип метода оплаты (опционально)
          example: "credit_card"
      required:
        - payment_id
        - user_id
        - amount
        - status
        - timestamp
    EventResponse:
      type: object
      properties:
        status:
          type: string
          description: Статус операции
          example: "success"
        event_id:
          type: string
          description: Уникальный идентификатор опубликованного события
          example: "38db80f04c9543049222cfef45ccbb11"
        topic:
          type: string
          description: Kafka topic, в который опубликовано событие
          example: "movie-events"
      required:
        - status
        - event_id
        - topic
    Error:
      type: object
      properties:
        error:
          type: string
          description: Сообщение об ошибке
          example: "Internal Server Error"
      required:
        - error